Update episode int to float

+++ src/AnimeDl/Models/Episode.cs
@@ -7,7 +7,7 @@ namespace AnimeDl
         public int Id { get; set; }
         public int AnimeId { get; set; }
         public string EpisodeName { get; set; }
-        public int EpisodeNumber { get; set; }
+        public float EpisodeNumber { get; set; }
         public string EpisodeLink { get; set; }
         public DateTime? CreatedAt { get; set; }
         public DateTime? UpdatedAt { get; set; }
